requirements.txt
PySide6
keyboard
mss
Pillow
numpy
manga_ocr
transformers
googletrans==4.0.0rc1
pykakasi
gTTS
sounddevice
soundfile
google-cloud-texttospeech
google-oauth2